[PyTorch] Stringize kernel? names consistently during macro expansion
I found 2 instances where the NAME parameter passed in to the dispatch macros is not a C++ string (constant) (i.e. double quoted compile time string).

In one instance it is a single quoted multi-character constant (I don't know what this resolves to in practice), and in the other instance, it is an unquoted identified generated as a result of concatenating 2 identifiers using the `##` operator.

Differential Revision: [D24211393](https://our.internmc.facebook.com/intern/diff/D24211393/)

ghstack-source-id: 113928208
Pull Request resolved: https://github.com/pytorch/pytorch/pull/46074