[JIT] Add reference semantics to TorchScript classes
**Summary**
This commit adds reference semantics to TorchScript class types;
modifications made to them within TorchScript will be visible in Python.

**Test Plan**
This commit adds a unit test to `TestClassType` that checks that
modifications made to a class type instance passed into TorchScript are
visible in Python after executing the scripted function or module.

**Fixes**
This commit closes #41421.

ghstack-source-id: d4c0042388facef79c3c331d7b04f8c419d59b5a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/44324