[StaticRuntime] Add a 'speed of light' benchmark.
This PR adds a hand optimized version of DeepAndWide model with the goal
of estimating overheads of static runtime. While static runtime is
currently much faster than the existing JIT interpreter, it would be
useful to understand how close we are to an absolutely 0-overhead
system. Currently, this "ideal" implementation is 1.5x faster than the
static runtime on batchsize=1.

PS: The implementation could probably be optimized even more.

ghstack-source-id: 5b8bfd0ddfd26f4f31eb6f392000f3fe38f3368f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/46308